refactor: adapt imports (treewide)
