+++ frappe/desk/doctype/bulk_update/test_bulk_update.py
@@ -6,8 +6,7 @@ import time
 import frappe
 from frappe.core.doctype.doctype.test_doctype import new_doctype
 from frappe.desk.doctype.bulk_update.bulk_update import submit_cancel_or_update_docs
-from frappe.tests import IntegrationTestCase, UnitTestCase
-from frappe.tests.utils import timeout
+from frappe.tests import IntegrationTestCase, UnitTestCase, timeout
 
 
 class UnitTestBulkUpdate(UnitTestCase):
